src/network_verifier/data_layer/config_validator.py
"""
Configuration validator for network device configurations.
"""

from typing import Dict, List, Optional
from batfish.client.session import Session
from batfish.datamodel import *
from ..config import settings

class ConfigValidator:
    """Validates network configuration files and their contents."""
    
    def __init__(self):
        self.session = Session(
            host=settings.batfish_host,
            port=settings.batfish_port
        )
    
    def validate_snapshot(self, snapshot_name: str) -> Dict:
        """
        Validate a network snapshot using Batfish.
        
        Args:
            snapshot_name: Name of the snapshot to validate
            
        Returns:
            Dictionary containing validation results
        """
        self.session.set_snapshot(snapshot_name)
        
        validation_results = {
            'parse_warnings': self._check_parse_warnings(),
            'reference_check': self._check_references(),
            'undefined_references': self._check_undefined_references(),
            'unused_structures': self._check_unused_structures()
        }
        
        return validation_results
    
    def _check_parse_warnings(self) -> List[Dict]:
        """Check for parse warnings in the configuration."""
        parse_warnings = self.session.q.parseWarning(
            nodes=NodeSpecifier()
        ).answer().frame()
        return parse_warnings.to_dict('records')
    
    def _check_references(self) -> List[Dict]:
        """Check for reference issues in the configuration."""
        references = self.session.q.referenceCheck(
            nodes=NodeSpecifier()
        ).answer().frame()
        return references.to_dict('records')
    
    def _check_undefined_references(self) -> List[Dict]:
        """Check for undefined references in the configuration."""
        undefined = self.session.q.undefinedReferences(
            nodes=NodeSpecifier()
        ).answer().frame()
        return undefined.to_dict('records')
    
    def _check_unused_structures(self) -> List[Dict]:
        """Check for unused structures in the configuration."""
        unused = self.session.q.unusedStructures(
            nodes=NodeSpecifier()
        ).answer().frame()
        return unused.to_dict('records')
    
    def validate_config_elements(self, config_elements: Dict) -> Dict:
        """
        Validate specific configuration elements.
        
        Args:
            config_elements: Dictionary of configuration elements to validate
            
        Returns:
            Dictionary containing validation results for each element
        """
        validation_results = {}
        
        # Validate interfaces
        if 'interfaces' in config_elements:
            validation_results['interfaces'] = self._validate_interfaces(
                config_elements['interfaces']
            )
            
        # Validate routing configuration
        if 'routing' in config_elements:
            validation_results['routing'] = self._validate_routing(
                config_elements['routing']
            )
            
        # Validate ACLs
        if 'acls' in config_elements:
            validation_results['acls'] = self._validate_acls(
                config_elements['acls']
            )
            
        # Validate BGP configuration
        if 'bgp' in config_elements:
            validation_results['bgp'] = self._validate_bgp(
                config_elements['bgp']
            )
            
        return validation_results
    
    def _validate_interfaces(self, interfaces: List[Dict]) -> Dict:
        """Validate interface configurations."""
        # Add specific interface validation logic here
        return {'status': 'valid', 'issues': []}
    
    def _validate_routing(self, routing: Dict) -> Dict:
        """Validate routing protocol configurations."""
        # Add specific routing validation logic here
        return {'status': 'valid', 'issues': []}
    
    def _validate_acls(self, acls: List[Dict]) -> Dict:
        """Validate access control list configurations."""
        # Add specific ACL validation logic here
        return {'status': 'valid', 'issues': []}
    
    def _validate_bgp(self, bgp: Dict) -> Dict:
        """Validate BGP configuration."""
        # Add specific BGP validation logic here
        return {'status': 'valid', 'issues': []} 