requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.0.1
aiofiles>=0.7.0
networkx>=2.6.3
pybatfish>=2024.7.22.1569
black>=21.9b0
flake8>=3.9.2
mypy>=0.910
pydantic>=2.0.0
python-dotenv>=0.19.0
pytest>=6.2.5 